refactor: step1 startLottoGame console 폴더로 이동
