+++ src/console/view/retryUntilValid.js
